Update patch set 1
Patch Set 1: -Code-Review

Well, I'm waiting on what direction everyone wants to take the overlays for the m7. Currently, this calls a non existent directory. But will it stay non existent in the near future? Are all the overlays for devices going to stay in the common, since they seem to be largely dependent on whether it's CDMA or GSM? For instance, I have an overlay for the m7vzw which fixes the roaming indicator being always on, even on vzw network. Should I try to move that to common, and see if the other devices don't break with it?

https://github.com/omnirom/android_device_htc_m7vzw/tree/android-4.4/overlay

Patch-set: 1
Reviewer: Gerrit User 1000453 <1000453@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0